requirements.txt
django
djangorestframework
django-cors-headers
django-filter
axios